Don't compile conntracker on non-linux platform
