+++ collector/pkg/metadata/conntracker/conntracker_others.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package conntracker
+
+func NewConntracker(config *Config) (Conntracker, error) {
+	return NewNoopConntracker(config), nil
+}
